Added generic implementations of CodeMap
The type of the positional representation is now parameterised.
All the behaviour and structs now have a second copy in the
optionally-included "generic" feature module.
This allows for codemaps to avoid the 4GiB memory limit imposed by
being restricted to u32. It might even be possible to use arbitrary-size
integers, but that may be limited by `usize`.